Emergency_Services_ALGOSSIAN
This gives the IDEA about the Emergency Services that are available in time of need or the emergency situation. 
It will give the nearest multi-speciality or specialised hospital according to the ratings and reviews of the earlier people those who have visited.
Ambulance contact, real time location and all the important factors related to that are taken care of with GPS.
Each of the important factors are taken care of and the things will be updated time to time in each phase with new updates and upgrades.
It is available in HINDI, ENGLISH and THE REIGONAL LANGUAGE the user selects while during registration 
A really great and beneficial support system that is made available.